requirements.txt
astunparse==1.6.3
blobfile==2.0.2
cupy
fastrlock==0.8
filelock==3.12.2
jittor==1.3.8.5
lxml==4.9.3
numpy
Pillow==10.0.0
pycryptodomex==3.18.0
six
tqdm==4.66.1
urllib3==2.0.4
